requirements.txt
Flask==2.0.1
flask-cors==3.0.10
google-cloud-logging==2.6.0